PP-13733 Move extract date & time functions to separate file
